Fixed bug where app would go into pip mode when player isn't active

+++ app/src/main/java/me/mauricee/pontoon/main/Player.kt
@@ -200,9 +200,11 @@ class Player @Inject constructor(preferences: Preferences,
     }
 
     override fun onPause() {
-        exoPlayer.playWhenReady = false
-        state = PlaybackStateCompat.STATE_PAUSED
-        focusManager.drop()
+        if (isActive()) {
+            exoPlayer.playWhenReady = false
+            state = PlaybackStateCompat.STATE_PAUSED
+            focusManager.drop()
+        }
     }
 
     override fun onStop() {
